split views into multiple files
makes it a bit more manageable.
